+++ gha_utils/labels.py
@@ -29,13 +29,8 @@ Files available:
 
 from __future__ import annotations
 
-import sys
+from importlib.resources import as_file, files
 from pathlib import Path
-
-if sys.version_info >= (3, 9):
-    from importlib.resources import as_file, files
-else:
-    from importlib_resources import as_file, files  # type: ignore[import-not-found]
 
 
 def _get_data_path(filename: str) -> Path: